refactor: phase 4 architecture — ConversationStore split, deferred parsing, consistency
- Split ConversationStore into MessageStore + ToolStatusStore sub-stores
- Add deferred parsing to MarkdownDocument (markdown_vault) — batch edits reparse once
- Audit tool return types: wrap error paths in ToolResult consistently
- Add explanatory comments to magic numbers (embeddings, memory_context)
- Extract checkbox constants in todos.py
